infrastructure/requirements.txt
aws-cdk-lib==2.162.1
constructs>=10.0.0,<11.0.0
boto3==1.36.9
aws-cdk.lambda-layer-kubectl-v31==2.0.2